+++ src/proxy-ui-api/src/test/java/org/niis/xroad/restapi/service/TokenCertificateServiceTest.java
@@ -25,10 +25,10 @@
 package org.niis.xroad.restapi.service;
 
 import ee.ria.xroad.common.CodedException;
+import ee.ria.xroad.signer.protocol.dto.CertRequestInfo;
 import ee.ria.xroad.signer.protocol.dto.CertificateInfo;
 import ee.ria.xroad.signer.protocol.dto.KeyInfo;
 import ee.ria.xroad.signer.protocol.dto.KeyUsageInfo;
-import ee.ria.xroad.signer.protocol.dto.TokenInfo;
 
 import lombok.extern.slf4j.Slf4j;
 import org.junit.Before;
@@ -38,9 +38,9 @@ import org.niis.xroad.restapi.exceptions.ErrorDeviation;
 import org.niis.xroad.restapi.facade.GlobalConfFacade;
 import org.niis.xroad.restapi.facade.SignerProxyFacade;
 import org.niis.xroad.restapi.repository.ClientRepository;
+import org.niis.xroad.restapi.service.TokenCertificateService.CsrNotFoundException;
 import org.niis.xroad.restapi.util.CertificateTestUtils.CertificateInfoBuilder;
 import org.niis.xroad.restapi.util.TokenTestUtils;
-import org.niis.xroad.restapi.util.TokenTestUtils.KeyInfoBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -52,6 +52,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import static ee.ria.xroad.common.ErrorCodes.SIGNER_X;
 import static ee.ria.xroad.common.ErrorCodes.X_CERT_NOT_FOUND;
+import static ee.ria.xroad.common.ErrorCodes.X_CSR_NOT_FOUND;
 import static ee.ria.xroad.common.ErrorCodes.X_INTERNAL_ERROR;
 import static ee.ria.xroad.common.ErrorCodes.X_TOKEN_NOT_AVAILABLE;
 import static ee.ria.xroad.common.ErrorCodes.X_TOKEN_READONLY;
@@ -73,12 +74,17 @@ import static org.mockito.Mockito.verify;
 @WithMockUser
 public class TokenCertificateServiceTest {
 
-    // cert hashes for mocking
+    // hashes and ids for mocking
     private static final String EXISTING_CERT_HASH = "ok-cert";
     private static final String EXISTING_CERT_IN_AUTH_KEY_HASH = "ok-cert-auth";
     private static final String EXISTING_CERT_IN_SIGN_KEY_HASH = "ok-cert-sign";
+    private static final String GOOD_KEY_ID = "key-which-exists";
+    private static final String KEY_NOT_FOUND_KEY_ID = "key-404";
     private static final String AUTH_KEY_ID = "auth-key";
     private static final String SIGN_KEY_ID = "sign-key";
+    private static final String GOOD_CSR_ID = "csr-which-exists";
+    private static final String CSR_NOT_FOUND_CSR_ID = "csr-404";
+    private static final String SIGNER_EXCEPTION_CSR_ID = "signer-ex-csr";
 
     // for this signerProxy.getCertForHash throws not found
     private static final String NOT_FOUND_CERT_HASH = "not-found-cert";
@@ -108,24 +114,58 @@ public class TokenCertificateServiceTest {
 
     @Before
     public void setUp() throws Exception {
-        // attempts to delete either succeed or throw specific exceptions
+        // need lots of mocking
+        // construct some test keys, with csrs and certs
+        // make used finders return data from these items:
+        // keyService.getKey, signerProxyFacade.getKeyIdForCerthash,
+        // signerProxyFacade.getCertForHash
+        // mock delete-operations (deleteCertificate, deleteCsr)
+        CertRequestInfo goodCsr = new CertRequestInfo(GOOD_CSR_ID, null, null);
+        CertRequestInfo signerExceptionCsr = new CertRequestInfo(
+                SIGNER_EXCEPTION_CSR_ID, null, null);
+        CertificateInfo authCert = new CertificateInfoBuilder().id(EXISTING_CERT_IN_AUTH_KEY_HASH).build();
+        CertificateInfo signCert = new CertificateInfoBuilder().id(EXISTING_CERT_IN_SIGN_KEY_HASH).build();
+        KeyInfo authKey = new TokenTestUtils.KeyInfoBuilder().id(AUTH_KEY_ID)
+                .keyUsageInfo(KeyUsageInfo.AUTHENTICATION)
+                .csr(goodCsr)
+                .cert(authCert)
+                .build();
+        KeyInfo goodKey = new TokenTestUtils.KeyInfoBuilder()
+                .id(GOOD_KEY_ID)
+                .csr(goodCsr)
+                .csr(signerExceptionCsr)
+                .build();
+        KeyInfo signKey = new TokenTestUtils.KeyInfoBuilder()
+                .id(SIGN_KEY_ID)
+                .keyUsageInfo(KeyUsageInfo.SIGNING)
+                .csr(goodCsr)
+                .cert(signCert)
+                .build();
+
+        // keyService.getKey(keyId)
+        doAnswer(invocation -> {
+            String keyId = (String) invocation.getArguments()[0];
+            switch (keyId) {
+                case AUTH_KEY_ID:
+                    return authKey;
+                case SIGN_KEY_ID:
+                    return signKey;
+                case GOOD_KEY_ID:
+                    return goodKey;
+                default:
+                    throw new KeyNotFoundException("unknown keyId: " + keyId);
+            }
+        }).when(keyService).getKey(any());
+
+        // signerProxyFacade.getKeyIdForCerthash(hash)
         doAnswer(invocation -> {
             String certHash = (String) invocation.getArguments()[0];
-            switch (certHash) {
-                case EXISTING_CERT_HASH:
-                    return null;
-                case SIGNER_EX_CERT_WITH_ID_NOT_FOUND_HASH:
-                    throw signerException(X_CERT_NOT_FOUND);
-                case SIGNER_EX_INTERNAL_ERROR_HASH:
-                    throw signerException(X_INTERNAL_ERROR);
-                case SIGNER_EX_TOKEN_NOT_AVAILABLE_HASH:
-                    throw signerException(X_TOKEN_NOT_AVAILABLE);
-                case SIGNER_EX_TOKEN_READONLY_HASH:
-                    throw signerException(X_TOKEN_READONLY);
-                default:
-                    throw new RuntimeException("bad switch option: " + certHash);
+            if (certHash.equals(EXISTING_CERT_IN_AUTH_KEY_HASH)) {
+                return AUTH_KEY_ID;
+            } else {
+                return SIGN_KEY_ID;
             }
-        }).when(signerProxyFacade).deleteCert(any());
+        }).when(signerProxyFacade).getKeyIdForCerthash(any());
 
         // signerProxyFacade.getCertForHash(hash)
         doAnswer(invocation -> {
@@ -147,46 +187,40 @@ public class TokenCertificateServiceTest {
             }
         }).when(signerProxyFacade).getCertForHash(any());
 
-        // construct a test auth key, sign key, and corresponding certs
-        // auth checks will use these keys to check from key whether auth or sign permissions are needed
-        TokenInfo tokenInfo = TokenTestUtils.createTestTokenInfo("fubar");
-        CertificateInfo authCert = new CertificateInfoBuilder().id(EXISTING_CERT_IN_AUTH_KEY_HASH).build();
-        CertificateInfo signCert = new CertificateInfoBuilder().id(EXISTING_CERT_IN_SIGN_KEY_HASH).build();
-        KeyInfo authKey = new KeyInfoBuilder()
-                .id(AUTH_KEY_ID)
-                .keyUsageInfo(KeyUsageInfo.AUTHENTICATION)
-                .cert(authCert)
-                .build();
-        KeyInfo signKey = new KeyInfoBuilder()
-                .id(SIGN_KEY_ID)
-                .keyUsageInfo(KeyUsageInfo.SIGNING)
-                .cert(signCert)
-                .build();
-        tokenInfo.getKeyInfo().add(authKey);
-        tokenInfo.getKeyInfo().add(signKey);
-
-        // keyService.getKey(keyId)
-        doAnswer(invocation -> {
-            String keyId = (String) invocation.getArguments()[0];
-            switch (keyId) {
-                case AUTH_KEY_ID:
-                    return authKey;
-                case SIGN_KEY_ID:
-                    return signKey;
-                default:
-                    throw new RuntimeException("unknown keyId: " + keyId);
-            }
-        }).when(keyService).getKey(any());
-        // signerProxyFacade.getKeyIdForCerthash(hash)
+        // attempts to delete either succeed or throw specific exceptions
         doAnswer(invocation -> {
             String certHash = (String) invocation.getArguments()[0];
-            if (certHash.equals(EXISTING_CERT_IN_AUTH_KEY_HASH)) {
-                return AUTH_KEY_ID;
-            } else {
-                return SIGN_KEY_ID;
+            switch (certHash) {
+                case EXISTING_CERT_HASH:
+                    return null;
+                case SIGNER_EX_CERT_WITH_ID_NOT_FOUND_HASH:
+                    throw signerException(X_CERT_NOT_FOUND);
+                case SIGNER_EX_INTERNAL_ERROR_HASH:
+                    throw signerException(X_INTERNAL_ERROR);
+                case SIGNER_EX_TOKEN_NOT_AVAILABLE_HASH:
+                    throw signerException(X_TOKEN_NOT_AVAILABLE);
+                case SIGNER_EX_TOKEN_READONLY_HASH:
+                    throw signerException(X_TOKEN_READONLY);
+                default:
+                    throw new RuntimeException("bad switch option: " + certHash);
             }
-        }).when(signerProxyFacade).getKeyIdForCerthash(any());
+        }).when(signerProxyFacade).deleteCert(any());
 
+        // signerProxyFacade.deleteCertRequest(id)
+        doAnswer(invocation -> {
+            String csrId = (String) invocation.getArguments()[0];
+            if (GOOD_CSR_ID.equals(csrId)) {
+                return null;
+            } else if (SIGNER_EXCEPTION_CSR_ID.equals(csrId)) {
+                throw CodedException.tr(X_CSR_NOT_FOUND,
+                        "csr_not_found", "Certificate request '%s' not found", csrId)
+                        .withPrefix(SIGNER_X);
+            } else if (CSR_NOT_FOUND_CSR_ID.equals(csrId)) {
+                throw new CsrNotFoundException("not found");
+            } else {
+                throw new CsrNotFoundException("not found");
+            }
+        }).when(signerProxyFacade).deleteCertRequest(any());
     }
 
     private CodedException signerException(String code) {
@@ -251,5 +285,40 @@ public class TokenCertificateServiceTest {
     public void deleteSignCertificateWithoutPermission() throws Exception {
         tokenCertificateService.deleteCertificate(EXISTING_CERT_IN_SIGN_KEY_HASH);
     }
+
+
+    @Test(expected = KeyNotFoundException.class)
+    @WithMockUser(authorities = { "DELETE_SIGN_CERT", "DELETE_AUTH_CERT" })
+    public void deleteCsrKeyNotFound() throws Exception {
+        tokenCertificateService.deleteCsr(KEY_NOT_FOUND_KEY_ID, GOOD_CSR_ID);
+    }
+    @Test(expected = CsrNotFoundException.class)
+    @WithMockUser(authorities = { "DELETE_SIGN_CERT", "DELETE_AUTH_CERT" })
+    public void deleteCsrCsrNotFound() throws Exception {
+        tokenCertificateService.deleteCsr(GOOD_KEY_ID, CSR_NOT_FOUND_CSR_ID);
+    }
+    @Test(expected = CsrNotFoundException.class)
+    @WithMockUser(authorities = { "DELETE_SIGN_CERT", "DELETE_AUTH_CERT" })
+    public void deleteCsrSignerExceptions() throws Exception {
+        tokenCertificateService.deleteCsr(GOOD_KEY_ID, SIGNER_EXCEPTION_CSR_ID);
+    }
+    @Test(expected = AccessDeniedException.class)
+    @WithMockUser(authorities = { "DELETE_SIGN_CERT" })
+    public void deleteAuthCsrWithoutPermission() throws Exception {
+        tokenCertificateService.deleteCsr(AUTH_KEY_ID, GOOD_CSR_ID);
+    }
+    @Test(expected = AccessDeniedException.class)
+    @WithMockUser(authorities = { "DELETE_AUTH_CERT" })
+    public void deleteSignCsrWithoutPermission() throws Exception {
+        tokenCertificateService.deleteCsr(SIGN_KEY_ID, GOOD_CSR_ID);
+    }
+    @Test
+    @WithMockUser(authorities = { "DELETE_SIGN_CERT", "DELETE_AUTH_CERT" })
+    public void deleteCsr() throws Exception {
+        // success
+        tokenCertificateService.deleteCsr(GOOD_KEY_ID, GOOD_CSR_ID);
+        verify(signerProxyFacade, times(1)).deleteCertRequest(GOOD_CSR_ID);
+    }
+
 
 }
